adding check to ensure that vendor module directories are initialized and not empty directories

+++ bridge/setup.sh
@@ -2,10 +2,12 @@ cd /usr/src/app
 
 for dir in /usr/src/vendor/* ; do
   if [[ -d $dir ]]; then
-    echo "Manually linking $dir"
-    dir_name=$(basename $dir)
-    rm -rf /usr/src/app/node_modules/$dir_name
-    cp -rp $dir /usr/src/app/node_modules/$dir_name
+    if [ "$(ls -A $dir)" ]; then
+      echo "Manually linking $dir"
+      dir_name=$(basename $dir)
+      rm -rf /usr/src/app/node_modules/$dir_name
+      cp -rp $dir /usr/src/app/node_modules/$dir_name
+    fi
   fi
 
   echo "Rebuilding linked modules"
